requirements.txt
# Data Analytics Pipeline Dependencies

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0

# Visualization & Reporting
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.0
